vent/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "jonathan-d-zhang", email = "69145546+jonathan-d-zhang@users.noreply.github.com"},
]
dependencies = [
    "Mastodon-py>=1.8.1",
    "pyzmq>=25.1.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

